6627055 nxge driver triggers bcopy error and panics system when setting kmem_flag in /etc/system (fix lint)

+++ usr/src/uts/common/io/nxge/nxge_txdma.c
@@ -313,7 +313,7 @@ nxge_fill_tx_hdr(p_mblk_t mp, boolean_t fill_len,
 	while ((nmp != NULL) && (tmp > 0)) {
 		size_t buflen;
 		mblk_len = MBLKL(nmp);
-		buflen = min(tmp, mblk_len);
+		buflen = min((size_t)tmp, mblk_len);
 		bcopy(nmp->b_rptr, cursor, buflen);
 		cursor += buflen;
 		tmp -= buflen;
